rename: 경기중계 컴포넌트 폴더 구조 정리 및 불필요한 컴포넌트 삭제
