+++ src/app/(route)/(community)/gameboard/_components/_baseballComponents/BaseballPlayerInfo.tsx
