+++ src/app/(route)/(community)/gameboard/_components/_baseballComponents/BaseballRecordItem.tsx
