+++ src/app/(route)/(community)/gameboard/_components/_footballComponents/FootballLivePanel.tsx
